part3_task/requirements.txt
pandas
numpy
psycopg2-binary
python-dotenv
supabase
schedule
pandas
python-dotenv
supabase